refactoring TasksPage and TasksProfilePage components
