+++ src/pages/tasks/styles.module.css
@@ -1,60 +1,6 @@
-.link {
-  text-decoration: none;
-}
-
-.tabContainer__text {
-  margin-right: 6px;
-  color: var(--colors-interface-additional);
-}
-
-.tabContainer__textActive {
-  color: var(--colors-interface-black);
-}
-
-.tabContainer__item {
-  display: flex;
-  margin-right: 40px;
-}
-
-.tabContainer__itemActive {
-  position: relative;
-  padding-bottom: 4px;
-}
-
-.tabContainer__itemActive::before {
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-  height: 2px;
-  background: var(--colors-interface-primary);
-  border-radius: 2px;
-  content: '';
-}
-
-.tabContainer {
-  display: flex;
-}
-
-.tabContainer__number {
-  background: var(--colors-interface-fuchsia);
-  border-radius: 9px;
-  min-width: 18px;
-  height: 18px;
-  color: var(--colors-interface-white);
-  text-align: center;
-}
-
 .input {
   max-width: 351px;
   margin: 30px 0 0 40px;
-}
-
-.users_card_box {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 20px;
-  margin-top: 20px;
 }
 
 .menu_block {
