+++ src/widgets/index.ts
@@ -13,7 +13,7 @@ export { ButtonWithModal } from './button-with-modal';
 export { InfoConflict } from './conflict-information';
 export { YandexMap } from './map';
 export { MapWithTasks } from './map-with-tasks';
-export { PageSubMenu, PageSubMenuForAdmins } from './page-sub-menu';
+export { PageSubMenu, PageSubMenuForAdmins, PageSubMenuLink } from './page-sub-menu';
 export { ProfileChatsPages } from './profile-chats';
 export { ModalContent } from './task-buttons-content';
 export { UserCard } from './user-card';
