+++ src/widgets/page-sub-menu/index.ts
@@ -1,2 +1,3 @@
 export { PageSubMenu } from './components/page-sub-menu/page-sub-menu';
 export { PageSubMenuForAdmins } from './components/page-sub-menu-for-admins';
+export { PageSubMenuLink } from './components/page-sub-menu-link/page-sub-menu-link';
